requirements.txt
mysql-connector-python
bandit